+++ spec/features/users_can_create_and_view_notes_spec.rb
@@ -8,7 +8,7 @@ RSpec.feature "Users can create and view notes" do
   let(:existing_note_created_at) {}
 
   before do
-    mock_successful_api_responses(urn: 12345, ukprn: 10061021)
+    mock_successful_api_responses(urn: 123456, ukprn: 10061021)
     sign_in_with_user(user)
 
     travel_to Date.yesterday do