+++ spec/features/users_can_view_the_actions_for_a_task_spec.rb
@@ -7,7 +7,7 @@ RSpec.feature "Users can view the Actions for a Task" do
   let(:project_id) { action.task.section.project.id }
 
   before do
-    mock_successful_api_responses(urn: 12345, ukprn: 10061021)
+    mock_successful_api_responses(urn: 123456, ukprn: 10061021)
 
     sign_in_with_user(user)
   end